Track register in Litmus CAS
In at least some cases (eg in C11 partialSC), R+W events have a register
associated with them, and that register feeds into the forbidden
execution.  For instance, one such event structure is (mind my ASCII):

[RMW x=0 -> x=1] --co-> [RMW x=0 -> x=2]

In this case, both RMWs store the read value to r0, because the
witnessing C execution is (0:r0 = 0 /\ 1:r0 = 0 /\ x=2).
However, at the moment, this register is thrown away at the Litmus
level and replaced with a temporary 'expected' value, causing the
postcondition to refer to a nonexistent register.

This is an attempt to push those registers through the Litmus IR and
into the pretty-printer for C.
I've assumed that there still exist possibilities where the RMW does
*not* spill to a register, so I've extended the Cas(...) variant with
an *optional* register, and kept the old logic around for when this is
None.